# docs/swagger/swagger.yaml
basePath: /api/v2
definitions:
...
  entity.CookedComment:
    properties:
      badge_color:
        type: string
      badge_name:
        type: string
      content:
        type: string
      content_marked:
        type: string
      date:
        type: string
      email_encrypted:
        type: string
      id:
        type: integer
      ip_region:
        type: string
      is_ai_assistant:
        type: boolean
      is_allow_reply:
        type: boolean
      is_collapsed:
        type: boolean
      is_pending:
        type: boolean
      is_pinned:
        type: boolean
      is_verified:
        type: boolean
      link:
        type: string
      nick:
        type: string
      page_key:
        type: string
      page_url:
        type: string
      rid:
        type: integer
      site_name:
        type: string
      ua:
        type: string
      user_id:
        type: integer
      visible:
        type: boolean
      vote_down:
        type: integer
      vote_up:
        type: integer
    required:
    - badge_color
    - badge_name
    - content
    - content_marked
    - date
    - email_encrypted
    - id
    - is_allow_reply
    - is_collapsed
    - is_pending
    - is_pinned
    - is_verified
    - link
    - nick
    - page_key
    - page_url
    - rid
    - site_name
    - ua
    - user_id
    - visible
    - vote_down
    - vote_up
    type: object
...
  entity.CookedPage:
    properties:
      admin_only:
        type: boolean
      date:
        type: string
      id:
        type: integer
      key:
        type: string
      pv:
        type: integer
      site_name:
        type: string
      title:
        type: string
      url:
        type: string
      vote_down:
        type: integer
      vote_up:
        type: integer
    required:
    - admin_only
    - date
    - id
    - key
    - pv
    - site_name
    - title
    - url
    - vote_down
    - vote_up
    type: object
...
  handler.Map:
    additionalProperties: true
    type: object
...
  handler.ParamsCommentCreate:
    properties:
      content:
        description: The comment content
        type: string
      email:
        description: The comment email
        type: string
      link:
        description: The comment link
        type: string
      name:
        description: The comment name
        type: string
      page_key:
        description: The comment page_key
        type: string
      page_title:
        description: The comment page_title
        type: string
      rid:
        description: The comment rid
        type: integer
      site_name:
        description: The site name of your content scope
        type: string
      ua:
        description: The comment ua
        type: string
    required:
    - content
    - email
    - name
    - page_key
    - site_name
    type: object
...
  handler.ResponseCommentCreate:
    properties:
      badge_color:
        type: string
      badge_name:
        type: string
      content:
        type: string
      content_marked:
        type: string
      date:
        type: string
      email_encrypted:
        type: string
      id:
        type: integer
      ip_region:
        type: string
      is_allow_reply:
        type: boolean
      is_collapsed:
        type: boolean
      is_pending:
        type: boolean
      is_pinned:
        type: boolean
      is_verified:
        type: boolean
      link:
        type: string
      nick:
        type: string
      page_key:
        type: string
      page_url:
        type: string
      rid:
        type: integer
      site_name:
        type: string
      ua:
        type: string
      user_id:
        type: integer
      visible:
        type: boolean
      vote_down:
        type: integer
      vote_up:
        type: integer
    required:
    - badge_color
    - badge_name
    - content
    - content_marked
    - date
    - email_encrypted
    - id
    - is_allow_reply
    - is_collapsed
    - is_pending
    - is_pinned
    - is_verified
    - link
    - nick
    - page_key
    - page_url
    - rid
    - site_name
    - ua
    - user_id
    - visible
    - vote_down
    - vote_up
    type: object
...
  handler.ResponseCommentList:
    properties:
      ai_comments:
        items:
          $ref: '#/definitions/entity.CookedComment'
        type: array
      comments:
        items:
          $ref: '#/definitions/entity.CookedComment'
        type: array
      count:
        type: integer
      page:
        $ref: '#/definitions/entity.CookedPage'
      roots_count:
        type: integer
    required:
    - comments
    - count
    - roots_count
    type: object
...
info:
  contact:
    email: artalkjs@gmail.com
    name: API Support
    url: https://artalk.js.org
  description: Artalk is a modern comment system based on Golang.
  license:
    name: MIT
    url: https://github.com/ArtalkJS/Artalk/blob/master/LICENSE
  title: Artalk API
  version: "2.0"
paths:
...
  /comments:
    get:
      consumes:
      - application/json
      description: Get a list of comments by some conditions
      operationId: GetComments
      parameters:
      - description: The user email
        in: query
        name: email
        type: string
      - description: Enable flat_mode
        in: query
        name: flat_mode
        type: boolean
      - description: The limit for pagination
        in: query
        name: limit
        type: integer
      - description: The username
        in: query
        name: name
        type: string
      - description: The offset for pagination
        in: query
        name: offset
        type: integer
      - description: The comment page_key
        in: query
        name: page_key
        required: true
        type: string
      - description: The scope of comments
        enum:
        - page
        - user
        - site
        in: query
        name: scope
        type: string
      - description: Search keywords
        in: query
        name: search
        type: string
      - description: Return AI assistant comments separately
        in: query
        name: split_ai
        type: boolean
      - description: The site name of your content scope
        in: query
        name: site_name
        type: string
      - description: Sort by condition
        enum:
        - date_asc
        - date_desc
        - vote
        in: query
        name: sort_by
        type: string
      - description: Message center show type
        enum:
        - all
        - mentions
        - mine
        - pending
        in: query
        name: type
        type: string
      - description: Only show comments by admin
        in: query
        name: view_only_admin
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseCommentList'
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Map'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get Comment List
      tags:
      - Comment
    post:
      consumes:
      - application/json
      description: Create a new comment
      operationId: CreateComment
      parameters:
      - description: The comment data
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/handler.ParamsCommentCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseCommentCreate'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Map'
            - properties:
                msg:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Map'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create Comment
      tags:
      - Comment
...
securityDefinitions:
  ApiKeyAuth:
    description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"